pyproject.toml
[project]
name = "gpu_monitor"
version = "0.1.0"
description = "GPU monitoring dashboard for distributed compute instances"
requires-python = ">=3.10"
dependencies = [
    "dash>=2.14.0",
    "dash-bootstrap-components>=1.5.0",
    "paramiko>=3.3.1",
    "pyyaml>=6.0.1",
    "pandas>=2.1.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
